Move augment local and augment grid methods to gk_geometry.c. Make a geometry input struct. Right now it is the same as the user side input struct. Will now make it hold more things so geometry function signatures can be shortened.
