+++ zero/gk_geometry.c
@@ -17,6 +17,97 @@ gkyl_gk_geometry_is_cu_dev(const struct gk_geometry* up)
   return GKYL_IS_CU_ALLOC(up->flags);
 }
 
+struct gkyl_rect_grid augment_grid(struct gkyl_rect_grid grid, struct gkyl_gyrokinetic_geometry_inp geometry)
+{
+  struct gkyl_rect_grid augmented_grid;
+  int cells[3];
+  double lower[3];
+  double upper[3];
+
+  if (grid.ndim==1) {
+    cells[0] = 1;
+    cells[1] = 1;
+    cells[2] = grid.cells[0];
+
+    lower[0] = geometry.world[0] - 1e-5;
+    lower[1] = geometry.world[1] - 1e-1;
+    lower[2] = grid.lower[0];
+
+    upper[0] = geometry.world[0] + 1e-5;
+    upper[1] = geometry.world[1] + 1e-1;
+    upper[2] = grid.upper[0];
+  }
+  else if (grid.ndim==2) {
+    cells[0] = grid.cells[0];
+    cells[1] = 1;
+    cells[2] = grid.cells[1];
+
+    lower[0] = grid.lower[0];
+    lower[1] = geometry.world[0] - 1e-1;
+    lower[2] = grid.lower[1];
+
+    upper[0] = grid.upper[0];
+    upper[1] = geometry.world[0] + 1e-1;
+    upper[2] = grid.upper[1];
+  }
+
+  gkyl_rect_grid_init(&augmented_grid, 3, lower, upper, cells);
+  return augmented_grid;
+}
+
+
+void augment_local(const struct gkyl_range *inrange,
+  const int *nghost, struct gkyl_range *ext_range, struct gkyl_range *range)
+{
+  if (inrange->ndim == 2) {
+    int lower_ext[GKYL_MAX_DIM], upper_ext[GKYL_MAX_DIM];
+    int lower[GKYL_MAX_DIM], upper[GKYL_MAX_DIM];
+    
+    lower_ext[0] = inrange->lower[0]-nghost[0];
+    upper_ext[0] = inrange->upper[0]+nghost[0];
+    lower[0] = inrange->lower[0];
+    upper[0] = inrange->upper[0];
+
+    lower_ext[1] = 1 - 1;
+    upper_ext[1] = 1 + 1;
+    lower[1] = 1;
+    upper[1] = 1;
+
+    lower_ext[2] = inrange->lower[1]-nghost[1];
+    upper_ext[2] = inrange->upper[1]+nghost[1];
+    lower[2] = inrange->lower[1];
+    upper[2] = inrange->upper[1];
+
+
+    gkyl_range_init(ext_range, inrange->ndim+1, lower_ext, upper_ext);
+    gkyl_sub_range_init(range, ext_range, lower, upper);  
+  }
+  else if (inrange->ndim == 1) {
+    int lower_ext[GKYL_MAX_DIM], upper_ext[GKYL_MAX_DIM];
+    int lower[GKYL_MAX_DIM], upper[GKYL_MAX_DIM];
+    
+    lower_ext[0] = 1 - 1;
+    upper_ext[0] = 1 + 1;
+    lower[0] = 1;
+    upper[0] = 1;
+
+    lower_ext[1] = 1 - 1;
+    upper_ext[1] = 1 + 1;
+    lower[1] = 1;
+    upper[1] = 1;
+
+    lower_ext[2] = inrange->lower[0]-nghost[0];
+    upper_ext[2] = inrange->upper[0]+nghost[0];
+    lower[2] = inrange->lower[0];
+    upper[2] = inrange->upper[0];
+
+
+
+    gkyl_range_init(ext_range, inrange->ndim+2, lower_ext, upper_ext);
+    gkyl_sub_range_init(range, ext_range, lower, upper);  
+  }
+}
+
 struct gk_geometry*
 gkyl_gk_geometry_deflate(const struct gk_geometry* up_3d, const struct gkyl_rect_grid* grid, const struct gkyl_range *local, const struct gkyl_range* local_ext, 
   const struct gkyl_basis* basis, bool use_gpu)
